implement requested changes: - rename and move findStrategy.js, - change indexOf -> stratsWith, - change if -> ternary operator.
